Introducing the flinch-o-meter (an adrenaline simulator) for an enhancement to DH's suppression system. Basically, the more you get shot at in a short period the higher your Flinch Meter goes. The higher the Flinch Meter the less involuntary flinches will affect you. However, the higher the Flinch Meter the more weapon sway you'll experience.
Why was this created?
Flinching is very effective. In a common example we see in DH, a SMG user is shooting at a Rifle user. The rifle in 98% of cases will not be able to kill the SMG user while receiving constant flinches. We feel with enough skill/experience a soldier would and could be able to sustain their flinches, as they become less of a reaction and more expected. This system is put in place to simulate that.

Why more sway?
This change would otherwise lesson the effect of suppression to a point we would not be happy with. By swapping the "flinching" for the "sway" we effectively make returning fire under suppression more skill based. Dealing with more sway can be overcome with player skill and timing, but flinches are basically luck.

Won't this nerf volume fire (lots of suppressive fire)?
Depends on the player being suppressed, however it will still be effective initially. We can easily adjust how effective initial suppression is. We can also adjust how many flinches is required before flinches have no effect (and vice versa with sway).

Do snaps and whizzes still blur the victim's view?
Yes and with full force, even once flinches have no effect and increased sway is in full effect.

This is still a WORK IN PROGRESS