Progress on Satchels and a big change to obstacles:  * Satchels can now consistently destroy obstacles, even if the obstacle's origin is under the world or in geometry  * There is now default obstacle's which every level with obstacles will use, this allows us to make changes to an obstacle without having to update every map  * ^ this system should not harm any level, including those with custom obstacles setup  * Satchels can now also consistently damage constructions, but this is WIP and more work is needed
How the default obstacle system works:
The default obstacles take priority. A level's DHObstacleInfo cannot override default obstacles. This allows for DH to both have consistent behavior of obstacles and for us to make changes to their behavior without having to update every level. This basically means there are a lot of Types setup in levels that aren't doing anything and they can be removed.