Reverted `DH_GUI_Tex` package back to v9.4.5b
This is related to the crashing issue #1104. I haven't been able
to replicate the bug after reverting the package, but i can't say
for certain that this fixes it (the replication steps are very
unreliable).

Crashes do occur after reverting and repackaging the background image,
so there might be something shifty going on with the texture.

This reverts changes made in commits:
* d81d86f5c460b5ad6c336d5c49c0893b7d3e6f28
* 749a79efcef9dd41d1f5e3ba77be5cbf10bc724f