Resolved #1106
The inventory system fails to add both GL and the satchel, because
one of them is a child of another.

When `VerifyGivenItems()` is called during spawn, items in the inventory
don't match the ones defined in the role info. That results in
`bRecievedInitialLoadout` left unset, which makes switching weapons
impossible.