Re-factor of spotting scope code:
There is now a "spotting scope" position index that can be set. This will display a scope that is fixed to the yaw of the gun, with no pitching (like a proper spotting scope!) This will allow for significantly easier additions of similar positions for different artillery pieces.

Also did a clean-up and refactor of the yaw knob drawing code to rely less on magic numbers. Still not complete yet (pitch rendering remains the same, for example), but it's a good start.