adjustment of recoil patterns on fast auto weapons:
In one of recent updates i discovered that there is an inconsistency in recoil between "fast auto" weapons and normal "auto", which seems to be caused by the special double shot mechanic that exists in "fast auto" that seemingly always makes the first shot into "double" (practically impossible to actually single tap), thus also applying double recoil for what appears to be the first shot, causing the weapon to jump higher after the first shot sound than if the weapon was just regular auto.

This makes it harder to control short bursts on "fast auto" weapons than on "auto", even if their recoil values are the same. This became very apparent to me after PPS-43 was switched from "fast auto" to normal, it became significantly more controllable and consistent in short bursts without any changes to recoil values themselves.

In order to reduce this recoil effect from first 2 shots being "merged", thus causing the weapon to "jump" higher in the first moment of the burst, I decided to lower the recoil for the first 1-2 shots on the recoil curve of "fast auto" weapons, while "restoring" a regular value on the second shot already so that this change only applies for the first 2 shots in the burst. Also in small compensation for this first shot "buff", recoil overall was increased a tiny bit.