[La Fiere Day Advance] Hand-tuned the Danger Zone
The lines were set to be tight and close to the objectives to prevent
both teams from aggressive flanking, especially in the first half
of the map.

This should also give Allies (albeit minor) protection for reinforcing
the Church area, as their only safe access to it is through the narrow
causeway.

* Enabled `bMainSpawn` on all main spawns.
* Set Danger Zone influences to:
    ```
    [0:La Fière]
    AxisInfluenceMoifier=0.00
    NeutralInfluenceMoifier=0.20
    [1:Estate Farm]
    AxisInfluenceMoifier=5.00
    NeutralInfluenceMoifier=0.10
    [2:Church]
    AxisInfluenceMoifier=6.00
    NeutralInfluenceMoifier=0.50
    [3:Causeway]
    AlliesInfluenceMoifier=2.00
    [4:Barn]
    [5:La Fière Manor]
    AxisInfluenceMoifier=4.00
    [6:Exit]
    AlliesInfluenceMoifier=5.00
    NeutralInfluenceMoifier=0.10
    ```