Removed watchtowers from the list of available constructions.
The general consensus is that they are useless in the role we intended for them (high-point observation and firing positions), and are instead useful only in their ability to enable map exploits.

It was fun while it lasted, so long friend.

The watchtower construction is still available to be pre-placed by levelers.