Added a new system that will invalidate the player's spawn point selection upon death if a more desirable spawn point becomes available.
The list of spawn point desirability is listed below, in descending order of desirability:

* Paradrop
* Squad Rally Point
* Platoon HQ
* Objective Spawn
* Progression Spawn (Push-style spawn points)
* Main Spawn

For example, if the player has a Platoon HQ selected as their spawn point and dies, but a Squad Rally Point is available, the player's spawn selection will be invalidated and the they will be prompted to select their spawn point.

This generally allows for better spawn option discovery (especially important for new players!), and also eliminates the common annoyance of accidentally spawning back at main spawn when better options become available throughout the match.