Squad leader can now use shovels near squadmates
Squad leaders can now use shovels as long as they have a squadmate
within 25 meters of them. This restriction should still severely limit
the prevalence of lone-wolf Minecraft squad leaders while still allowing
the shovel to be used for destruction of enemy assets and building
assets so long as there are squadmates nearby.