ASLs no longer auto-assigned SL if SL leaves squad
The ASL will also be omitted from being automatically selected as the
new squad leader if no one volunteers to be the squad leader.

This is done since assistant squad leaders are often designated the
primary builder, and if they are given SL, they may suddenly lose the
ability to build.

An ASL can still volunteer to become the squad leader, at which point,
they have volunteered to limit their ability to build constructions.