Constructions can be torn down when SL is demoted
Unbuilt constructions placed by a squad leader that is demoted (by
leaving or losing the SL role for any reason) will automatically be torn
down and will have their supplies refunded.

This is to stop the potential exploit where players could place a bunch
of constructions as SL, then drop SL and be able to build them solo.