Soviet vehicles coax reload buff 5.5 seconds
-Coax foley changed
-Coax reload speed improved to 5.5 seconds
-Coax on the t34/76 m41 is 6.5 seconds due to poor ergonomics

Soviet vehicles were notorious for being hard to work with when it comes to infantry support, the coax was on average far longer to reload than its belt fed contemporaries 7.6 seconds at its longest compared to the 6 odd seconds of the belt feds.

Coax reload speed improved to 5.5 seconds, reasoning being its a pan mag that is easily hit up out of the way and then put into place, as compared to the whole feeding process of belt fed machineguns.