Mortars can now deploy in "no arty zones".
This should relieve the irritating problem of not being able to deploy
mortars in a team's starting zone.

This makes mortars immune to counter-battery fire if they are within a
no-arty zone, but I don't think this is particularly bad. If a player
can spot an enemy mortar for counter-battery fire, they can probably
just as easily shoot the guy themselves.