Recoil curve added to all full-power semi autos
Recoil curve added to every semi-auto with a full power cartridge, recoil curve tweaked between the Garand and others.

Garand still cannot be slam fired and will jump off target, but not to the extent currently. however, the other semi autos are easier to control due to their blade sights.