Fix:  Some Soviet tanks do not start with their cannon loaded
This was a bug caused by a mismatch in redundant class property settings ProjectileClass and PrimaryProjectileClass. There is no reason why these two should ever be different when a tank is first spawned, so we now simply enforce this in PostBeginPlay of DHVehicleCannon.

All the redundant default property lines have now also been removed.