Fix: Vehicle Pools spawn kill behavior is broken
This fixes a bug where vehicles that were spawn killed on vehicle pools
would not be treated as spawn kills and subsequently not destroy the
vehicle pool itself, allowing players to endlessly camp vehicle pools.

This was a result of two things:

1. The spawn kill timer was using the wrong variable (it was using the
spawn protection timer, not the spawn kill timer).
2. The timer it was using was only 2 seconds before, making it virtually
impossible to trigger this behavior. It has now been boosted to 15
seconds.

This makes vehicle pools more susceptible to being destroyed by proxy,
but dramatically reduces the likelihood of vehicle pool spawn camping.