Gameplay: Adjusted spawn point hint behavior
This changes how spawn points use spawn hints to spawn players into the
level.

Before, the hint locations were chosen effectively at random every time,
making it technically possible to spawn players and vehicles inside of
one another if the pawn intersection test failed (as it sometimes does).

The new scheme will use spawn hints in a predictable order to reduce the
likelihood of people spawning inside of each other.

For example, if there are 4 spawn hints (A, B, C, and D), the order that
the spawn hints will be used is:

A, B, C, D, A, B, C, etc...

The progression of the sequence is tracked separately for both infantry
and vehicle hints.