Breda first-person animations finalized
The reload animation now works for any number of bullets remaining in the magazine (using a specialized system of animation notifies).

"Floating rounds" in the feed chamber will now also spill out of the gun on reload. This has no effect on the number of bullets left, it's just a visual effect.