Squad assistants are not immediately assigned to be the new squad leader after the squad leader leaves.
ASLs will also not be eligible to be auto-assigned to SL if they did not explicitly volunteer to be SL.

The rationale for this is that in the current meta, ASLs are usually running logistics, and having an ASL automatically being turned into the SL means that they can no longer build without someone nearby.

In the future when we have a dedicated logistics squad, this whole thing can be ripped out as we figure out what the ASLs new job will be.