Vehicles: Changed satchel & engine fire mechanics
Satchels that explode on the tops of tanks and within range of the
engine will no longer 100% start a fire in the engine. The likelihood of
an engine fire is now determined by how close the explosion is to the
engine.

Furthermore, the likelihood of an engine fire spreading to the hull has
been dramatically reduced. In reality, most vehicles had firewalls that
separated the engine from the crew cabin, making this type of thing
unlikely.

Previously, due to some wonky programming, the likelihood of the hull
catching fire was virtually 100% (and ironically, made worse if your
engine had a lot of HP), since it was doing a fire-spread dice roll
every second while the engine was on fire. This system makes it
to difficult to reason about and balance the likelihood of fires for
different vehicles because the probabilities have to be amortized
in relation to the vehicle's HP which is far too much complexity for
what should be a simple dice roll.

These two changes together make it so that satchels are no longer a
guaranteed one-hit kill weapon against every tank in the game.