Increased M1919A6 and MG42 belt size from 150 to 200.
This is a partial reversion to the previous belt size a few months ago.
Players complained that the belt size was too small.

The tentative plan is to revert it back to 150 when we add heavy MGs,
since those will have huge belts, and the tradeoff of a stationary vs.
mobile MG will make the smaller belt size make sense.