Fixed an exploit where it was possible to place constructions in invalid locations
Players had been exploiting this for years without telling us the repro
steps. Luckily, we had an informant tell us how to reproduce it.

The reproduction steps were as follows:

1. Bring up the construction menu then select an AT gun.
2. While the proxy object is still active, bring up the construction
menu again, then choose a smaller construction.
3. Click to place the smaller construction.
4. Observe that the original construction (the AT gun) is placed.

This has been the root cause of a lot of construction oddities seen
over the years. Is this what it means...to make DH great again?