Refactored Command Post data model
Each nation-specific command post variant is now its own class. This
is much easier to manage than spreading the data into the nation class
where it doesn't belong.

This is made possible by a new `GetConstructionClass` function in
`DHConstruction` that lets a construction class act as a proxy for an
underlying construction class that is determined by the context (i.e.,
team, nation etc.). This allows us to have one DHConstruction_PlatoonHQ
class on the command wheel.