Fixed mortar animations being jittery and out of sync
Offending commit: 25bfbc278a1af20b28d72d8db83c1703aa816285

The bug was triggered by changing mortar anim rates and frame counts in
the character package.

The jitter has been fixed by removing an unnecessary `PlayAnim` call
from the `AnimEnd` event (which was only called for anims with more than
1 frame)

This was a holdover from RO code that currently has no apparent effects
other than breaking the mortars and causing a mess. Other vehicle
animations seem to be unaffected by the removal.

I have no idea what was the intention behind it, but whatever it was
trying to solve, this was a terrible way to do it.

Changes to `DHCharactersGER_anm`:

* Reverted the rates for `deploy_fire`, `flinch_return`, and `deploy`
  animations to values before commit
  25bfbc278a1af20b28d72d8db83c1703aa816285.

* Lowered the rates for `deploy_idle` animations to 0.2 for performance
  reasons (this prevents triggering `AnimEnd` too often).