Semovente 47/32 projectile direction is now aligned to the muzzle rotation
There was a systemic issue where the direction vector for projectiles
from tank cannons was being incorrectly calculated, made especially
noticable if the yaw axis of the cannon was not axis-aligned.

This behavior is preserved for backwards compatibilty, but tanks can be
switched to the new behavior as needed. A new `ProjectileRotationMode`
in `DHVehicleCannon` determines which mode is used. The Semovente 47/32
uses the new mode.

It should also be noted that with the old system, the AddedPitch
calculation is wildly incorrect, because it merely adds the pitch in
orld-space, not cannon-space. This means that tanks that have a
significant roll in their rotation will have nonsensical trajectories
when paired with a high added pitch (due to gunsight ranging).