Added support for enemy-specific voice commands
This has now been implemented on the Italians. The US-specific insults
will now be used when facing off against the US. Support has been also
added for the USSR, but we have no maps that use it yet. As a fallback
it will only use voice lines that are non-specific.