Major refactor to how a team's construction list is defined and managed
There is a now a `DHConstructionLoadout` which enumerates a list of
constructions. Nations now have `DefaultConstructionLoadoutClass`
references which are prefab construction loadouts.

The lever can override either the entire loadout class or make
adjustments to the existing loadout in `DH_LevelInfo`.

Numerous other minor refactors in the blast radius of this change.