initial commit
asdsad

incremental commit

not much done

Incremental commit, fixed the K98 scope this time

Figured out how to do the alpha on the scopes correctly, it should all fall in place now

Disabling bUsesIronsightFOV

This change is here to adress the certain "annoying" aspect of using scoped weapons, it is found that it is offputting to people if the FOV zooms in along with the weapon.

as well as this, keeping the FOV the same both incorporates a realistic aspect, as well as gives the user of the scoped weapon a greater perhipheral vison for any dangerous elements.

thank alti for this change frfr

Revert "Disabling bUsesIronsightFOV"

This reverts commit d1e749e6a247658da4a8dd5a9ccfa784fc2a1528.

british and german scopes complete. removed added yaw from the scoped enfield

-Re-textured british and german scopes
-removed AddedYaw from the scoped enfield (was leftover from trying to fix it)

textures complete

all scope textures complete

scopes complete (for real this time)

Zeroed and completed the scope refactor

-re-zeroed all scoped weapons
-full implimentation of new scope textures
-removed herobrine

Fixed the 2d scope problems

transparency issues on the textures were causing issues with 2d scopes, implimented specific 2d scope textures to negate this problem and removed "todo" code comments