Disabling bUsesIronsightFOV
This change is here to adress the certain "annoying" aspect of using scoped weapons, it is found that it is offputting to people if the FOV zooms in along with the weapon.

as well as this, keeping the FOV the same both incorporates a realistic aspect, as well as gives the user of the scoped weapon a greater perhipheral vison for any dangerous elements.

thank alti for this change frfr