Fix #1731: Vehicle radios do not work in multiplayer
There were 2 things going wrong here; first, the team index was not
being set correctly on the server, so it was effectively a neutral
radio.

The other more pressing issue was that the radio was not being properly
positioned on the client. It was being positioned using RelativeLocation
and, while this works fine if it's a decorative client-side attachment,
it does not work correctly for radios. For some reason, attaching it
to a bone makes it replicate the position correctly every time (same
as is done on the Fiat 508CM).