Custom Bayonet functionality
the Bayonet for a specific weapon can now be specified by a static mesh, this means that early war bayonets can be attached to late war rifles and possibly vice versa.

A modular bayonet attachment system that allows either the legacy bayonet system, which draws the mesh from the animations package, and the custom, which draws from the static mesh package to both work if needed.

thanks basnett