The build's git information is now written into the build
The <major>.<minor>.<path> version is now written in `make-config.json`
in the mod directory.

The server browser now checks that the commit hash is the same to
determine if the server is compatible with the current version. This is
more accurate than just checking the top-level version.

The version is no longer stored in `DarkestHourGame`, so this has been
removed.