The Panzerfaust crate now holds 4 Panzerfausts
This restores the total amount to what was originally intended. The
reason it was reduced down to 2 was for balance reasons since there
was effectively no cap on the amount of these things that could be
spawned before.

However, with the new construction limiting system, we can constrain
the total amount of spawned Panzerfausts to a finite number. In this
case, we'll say the Germans can have 6 boxes by default, which is
24 Panzerfausts in total at a supply cost of 5100.

This can be adjusted as balance requires in the future.