Fix #1753: Mantling with certain weapons breaks player animations
Replaced 77 `jump_land` and `jump_takeoff` sequences with single-frame
animations for the following weapons:
 - Bazooka M9A1
 - Bazooka M1A1
 - Panzerschreck
 - PIAT
 - M1919
 - Breda 30
 - Beretta M1934
 - M39 (unused animation)

When looped, single-frame animations act differently from animations
with multiple frames. This change of behavior caused pawns to get stuck
in `Mantling` state on remote clients.

Would be nice to fix this in code as well, but it's too much to unravel
at the moment.