Increased resupply count of vehicle resupply constructions from 15 to 25
In playtest, 15 was much too low considering how much these cost to
place. In addition, the value was lowered in part as a response to the
AT gun spam that was happening pre-Italy update. Since that are now more
stringent limits on the number of guns that can be placed, the
justification for this value being so low disappears.