Change to how vehicle cannon resupply is performed
The previous resupply behavior was to iterate through each round type
and increment the amount by one. This meant that it could resupply as
many as 3 rounds, or as few as 1, even if there was still more room to
add 2 more. This makes no sense!

The new system now has a variable number of rounds it can give per
resupply interval (definable in the cannon's class). Up to that many
rounds will be distributed evenly among the round types that can still
be resupplied.