Fixed a bug where you could not interact with stationary weapons despite the button prompt
Also, once a player is out of a particular round type on a cannon (tank
cannon, mortar etc.), it will not just automatically switch to the next
round to be fired.

The rationale behind this is that the user will not necessarily notice
that they are out of ammo, and may start firing the wrong ammo type
which is undesireable. There is a new TODO item to add a button prompt
so that the user knows they need to switch to a new round type.