Adding font downscaling when using more square aspect ratios
Since the font selection is dependent on the Y resolution in relation
to the baseline aspect ratio, we downscale the "effective" Y resolution
for the lookup.

This fixes an issue where players using 1280x1024 and
other 4:3 aspect ratios would have seemingly enormous text because it
was effectively assuming their X resolution was 1820.