Changed capitalisation of class references & removed package qualifiers
Class references are now formatted with a capital C for `Class`. The reason for this change is that a `Class` is the type component of the object reference, and not an UnrealScript keyword.

In addition, redundant package qualifiers have been removed entirely, since the compiler does not actually need them in the vast majority of cases. This has the benefit of shrinking line & file sizes, and also allows ctrl+click navigation using the UnrealScript extension in VS code.