Fixing map deco grass textures
Went through the maps and swapped out which textures the deco grass uses.
Also checked if the myLevel vegetation was set up correctly.
renamed the grass texture in the texture package, since I forgot to change the name after duplicating another texture (there were 2 of the same in different groups)