Changed Mesh replication logic mesh for vehicle constructions
The client can determine the mesh to use by itself based on the vehicle class. Therefore, we don't replicate the mesh.

Previously, we used `LinkMesh` and relied on the native replication to get the client to display the correct the mesh. However, if the mesh was in a package marked with the `ServerSideOnly` flag, the client would display an empty mesh.

Because we need to keep the amount of non-`ServerSideOnly` packages down to an absolute minimum, we changed it so that the client and server set the meshes independently.