Fixed bug where vehicles could disappear when destroyed
This was because the package containing the static mesh was marked
as ServerSideOnly, so when the StaticMesh property was replicated to
clients, the client would fail to resolve the reference.

Because the client and the server both know what the destroyed static
mesh should be, we can set the SM ahead of time and just flip the
draw type.