LMG recoil adjustments
Recoil adjusted to bring more consistency. Among the most affected is the DP-27, which for some bizarre reason was very different between late and regular version (they were supposed to be identical). Breda recoil was also reduced, as it is rightfully one of the few advantages it should have over the others (low powered cartridge in a heavy machinegun). The rest are mostly minor adjustments