slightly lowered SRCM35 damage
slightly reduced srcm35 damage to make it more consistent with the other offensive grenades (stg24 and rg34); should also help with balance issues.

TO DO: i think the damage radius on this grenade is very questionable and should be changed, but i will leave it for the future global grenade rebalance that will hopefully happen when we set up the dependency between damage and TNT equivalent.