Vehicle icons are now drawn on top of ammo icons
A common complaint is that ammo stack icons would obscure the more important icons (i.e. AT guns). This adds a sorting variable for the classes. The list of visible attachments is sorted each frame. This isn't too bad because the list is quite small.

A nicer solution would be to maintain a list of the map icons on the client and keep it sorted through insertion sorts, then the sorting would only need to take place whenever the list is added to.