Rolling the font generation into the DH_Fonts class itself
This simplifies things considerably, since the user no longer has to go into the SDK to run the generation script.

Only downside is that it doesn't really let you do this with multiple packages unless it also generates another package entirely, which is within the realm of possibility!

I will want to add something to the make script that ensures the user has all of the specified fonts installed on their machine before running the build.

I think for the time being we'll probably want to keep the DH_Fonts package commited to the repository, and it'll be up to the maintainer to make sure the generated fonts are good.