Re-adjusted the font generation pipeline
It's not desirable to have all of the fonts in one file, nor to make it mandatory to have all the fonts installed in order to build the game. Therefore, I've separated the `DHFonts` class from the actual font assets, restoring the previous workflow where you need to run the script through the editor.

Going into the editor kind of sucks, and I have an idea of how we can get the best of both worlds, but it would take too long to to implement it. For now, this will work!

One remaining major inefficiency is that we are over-generating glyphs for some fonts. Some fonts don't need the entire glyph set since they are used for very specific things (i.e., big fat buttons on the main menu). This balloons the font size quite a bit, especially for CJK fonts.